requirements.txt
pytest~=8.0.2
pyzbar~=0.1.9
pdfplumber~=0.10.4
opencv-python~=4.9.0.80
aspose_pdf==24.2.0